move cryptography and tcp send/recv pipeline code into its own module
